requirements.txt
FastAPI
uvicorn
python-dotenv
pydantic
sqlalchemy
pytest
httpx